[MultiServer3 Super] - Initial update commit.
This commit implements the next big update known as the "Super" variant.

The fixes ranges from optimizations, to feature improvements, security fixes, docker fixes, and server maintenance.

This also introduces a new server known as the EdenServer for TDU applications.

Among the fixes, there is a new config versioning, which will now properly handle config changes when that is needed.

For a detailed list of what has changed, please refer to each individual files listed.