[ApacheNet/EndianTools/HomeTools] - Implements missing Sodium Objective and OHS bar entries, code sanity checks.
The code sanity checks aim to eliminate the .Result usage in the HomeTools library, these are deadlock prone according to the documentation, so we use a "sync" version of the code where it makes sense, and async accelerate speed critical parts only.

Also moves the endian Binary methods to the endian library, where it makes sense of course (while eliminating a duplicate class).