Super update, batch 2.
This commit pushes the 2nd batch of update for the Super branch.

This further continues the work on the EdenServer, fixes several aspects of the backend, renames the NetworkLibrary to a better representative name, updates the multispy server to use a newer version of minircd compatible with python 3, prepares the terrain for MAG work base.