[QuazalServer/MultiSocks/Horizon/MitmDNS] - Switches to a SQLITE database engine for all the user storage + EntityFramework migration fixup + BP login fixes and adds a missing hostname in the routes.txt file.
This update implements a real database system for respective servers (NOTE: Horizon cannot read the CID bans for now, will be fixed shortly), as well as a fix for the BP login path (requests news 8 for some odd reasons).

On top of this, the database migration for pomelo entity framework has been implemented via a AspNet project (and updated procedure in the DbContext file.

This also adds a missing dns entry in the listing.