[WebAPIService/MultiServerLibrary/ApacheNet] - Bugfixes and directory sorting for the web stuff. Copies the build script with GUI compile.
This commit applies some bugfixes related to SSL certificate not loading up without an internet connection, apachenet not having the right suffix for the direcotory sorting on a non-default port and duplicate project references across projects.

This also sort files and folders in the Web API Service project to prepare the micro-services system.