[NEEDS TESTING] - Super update, batch 3.
This commit pushes the 3rd batch of update for the Super branch.

The main features, ranges from various bug fixes and improvements across the server stack.

The biggest feature improvement of the commit, is the re-write of ApacheNet which is now significantly simplified and a lot easier to use for newcomers, as well as fixing MS1 legacy code with a newer equivalent (massive perf boost for API lookup).

While we are at it, I removed a lot of junk where it made no sense.