[MultiServerLibrary/WebAPIService/ApacheNet] - Implements client proxy and fixes the GeoLite updater speed. Updates the Jwplayer 4.3 with an incremental update, and get rids of the swfobject system.
The file hashing system was extremly slow, so we ditch it entierly and read all bytes at once instead for the GeoIP updater.

Some networks relies on proxies, so we now handle that when needed within the library itself. This also removes the HttpClient based methods, I hate this system even on modern NET revisions.