[PS2FloatLibrary] - Partial revert of the commit: https://github.com/GitHubProUser67/MultiServer3/commit/9d1a2ed5a2b6aca8a0e20acc2462d1cd16f7a076
The MIT license is challenging to follow due to the multiplier being innacurate compared to the PCSX2's solution.

However, we can keep our C# port of the multiplier and fast div/sqrt (slow method was implemented by me originally), since PCSX2 is GPLv3, however, where there is code related to it, the copyright notice is being added.

Sorry for the PDTools software, but currently this isn't possible to port to a MIT licensed project.