[DNSLibrary] - Fixes several pointer-based exploits and very tiny code lookup.
The dynamic class filler system had a severe problem, where not everywhere in the library code the size of the struct is verified, leading to exploits being possible in this manner. We replace all the code using it by a manual endian-based filler system.

The getbytes method however is safe but unused in our cases.

While we are at it, fixes a very few warnings.