[WebAPIService] - Truly fix Sodium2 racer leaderboards and also fixes a bug with OHS weekly leaderboards sorting.
We were skipping the score entity while doing the weekly check, we fix that.

Also reverts our previous fix for Sodium2, the real issue was the default score 0 we input now, so for that, we now have a list of banned projects for the filler.

Also as a small bonus, print the order of the batch entries for diganosis.