[CastleLibrary/EndianTools] - Fixes SNI domains parsing regression.
Some domains failed to be processed by our method due to an encoding mismatch while handling dots.

To fix this we forked the NETCORE2.1 api while modifying it to handle extra extensions.

We still fallback to our method for safety (we handle non TLS clients as well, but they should be supported anyway).

Due to this, several Span base endian readers were implemented as well.